refactor(chat): extract chat_manager/shared.rs (#292)
Moves the free helpers shared between the facade and the sub-manager
Deps wrappers (is_working, enrich_chat, enrich_and_emit,
apply_tuning_impl, the queued-ref helpers, build_history_session,
remap_history, now_ms) into their own submodule per the plan's Task 5.
